[core] fix cycles in common/utils
